[model] Support HunyuanMT1.5-1.8B and HunyuanMT1.5-7B templates
- Add hunyuan_mt1_5_1_8b and hunyuan_mt1_5_7b template types to LLMTemplateType
- Move hunyuan_moe and hunyuan template definitions from llm.py to tencent.py
